+++ ecommerce/courses/tests/test_publishers.py
@@ -130,7 +130,7 @@ class LMSPublisherTests(DiscoveryTestMixin, TestCase):
         expected = {
             'name': 'verified',
             'currency': 'USD',
-            'price': int(stock_record.price_excl_tax),
+            'price': float(stock_record.price_excl_tax),
             'sku': stock_record.partner_sku,
             'bulk_sku': None,
             'expires': None,
@@ -162,7 +162,7 @@ class LMSPublisherTests(DiscoveryTestMixin, TestCase):
         expected = {
             'name': expected_mode,
             'currency': 'USD',
-            'price': int(stock_record.price_excl_tax),
+            'price': float(stock_record.price_excl_tax),
             'sku': stock_record.partner_sku,
             'bulk_sku': None,
             'expires': None,
@@ -178,7 +178,7 @@ class LMSPublisherTests(DiscoveryTestMixin, TestCase):
         expected = {
             'name': 'verified',
             'currency': 'USD',
-            'price': int(stock_record.price_excl_tax),
+            'price': float(stock_record.price_excl_tax),
             'sku': stock_record.partner_sku,
             'bulk_sku': ec_stock_record.partner_sku,
             'expires': None,
